Move DownloadInfo to rocket package
