+++ app/src/main/java/org/mozilla/focus/components/DownloadCompleteReceiver.java
@@ -13,7 +13,7 @@ import android.net.Uri;
 import android.os.Environment;
 import android.text.TextUtils;
 
-import org.mozilla.focus.download.DownloadInfo;
+import org.mozilla.rocket.download.data.DownloadInfo;
 import org.mozilla.focus.download.DownloadInfoManager;
 import org.mozilla.focus.telemetry.TelemetryWrapper;
 import org.mozilla.threadutils.ThreadUtils;
