+++ app/src/main/java/org/mozilla/rocket/download/data/DownloadInfo.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.focus.download;
+package org.mozilla.rocket.download.data;
 
 import android.text.TextUtils;
